improve test methods, remove double methods in other class

+++ src/main/java/ru/java/practicum/filmorate/storage/LikesStorage.java
@@ -14,5 +14,5 @@ public interface LikesStorage {
 
     List<Film> getPopularFilms(int count);
 
-    int getLikesCountForFilm(Long filmId);
+    Long getLikesCountForFilm(Long filmId);
 }
